Test only the GMPLib with "Include BitCode" enabled
Got the error from app store:

We have discovered one or more issues with your recent delivery for
"Stage-Plato". To process your delivery, the following issues must be
corrected:

Invalid Bundle - The app cannot be processed because options not
allowed to be embedded in bitcode are detected in the submission. It is
likely that you are not building the app with the toolchain provided in
Xcode. Rebuild your entire app with the latest GM Xcode and submit the
app again.